requirements.txt
# Core
google-generativeai>=0.7.0
python-dotenv>=1.0.0

# API clients
praw>=7.7.1                   # Reddit
google-api-python-client>=2.0 # YouTube
instagrapi>=2.0.0             # Instagram
tweepy>=4.14.0                # X / Twitter
httpx>=0.27.0                 # Facebook (raw HTTP)

# Web server
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.0.0